% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R,
%   R/methods-ModularExperiment.R, R/methods-ReducedExperiment.R
\name{cbind,FactorisedExperiment-method}
\alias{cbind,FactorisedExperiment-method}
\alias{cbind,ModularExperiment-method}
\alias{cbind,ReducedExperiment-method}
\title{Combine ReducedExperiment objects by columns}
\usage{
\S4method{cbind}{FactorisedExperiment}(..., deparse.level = 1)

\S4method{cbind}{ModularExperiment}(..., deparse.level = 1)

\S4method{cbind}{ReducedExperiment}(..., deparse.level = 1)
}
\arguments{
\item{...}{\link[ReducedExperiment]{ReducedExperiment} objects.}
}
\description{
Comnbines \link[ReducedExperiment]{ReducedExperiment} objects by columns
(samples). Combining these objects in such a way assumes that feature-level
slots (e.g., loadings, module membership) are the same for all objects to
be concatenated. If they are not, an error is returned.
}
