% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R
\name{dendrogram,ModularExperiment-method}
\alias{dendrogram,ModularExperiment-method}
\title{Get the dendrogram stored in a ModularExperiment}
\usage{
\S4method{dendrogram}{ModularExperiment}(object)
}
\arguments{
\item{object}{\link[ReducedExperiment]{ModularExperiment} object.}
}
\description{
Get the dendrogram stored in a ModularExperiment
}
