% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ReducedExperiment.R
\name{dim,ReducedExperiment-method}
\alias{dim,ReducedExperiment-method}
\title{Get the dimensions of a Reducedexperiment object}
\usage{
\S4method{dim}{ReducedExperiment}(x)
}
\arguments{
\item{x}{\link[ReducedExperiment]{ReducedExperiment} object.}
}
\value{
Returns a named vector containing the dimensions of the samples,
features and reduced dimensions.
}
\description{
Get the dimensions of a Reducedexperiment object
}
