% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R,
%   R/methods-ModularExperiment.R, R/methods-ReducedExperiment.R
\name{.DollarNames.FactorisedExperiment}
\alias{.DollarNames.FactorisedExperiment}
\alias{.DollarNames.ModularExperiment}
\alias{dollar_names}
\alias{.DollarNames.ReducedExperiment}
\title{Command line completion for \code{$}}
\usage{
\method{.DollarNames}{FactorisedExperiment}(x, pattern = "")

\method{.DollarNames}{ModularExperiment}(x, pattern = "")

\method{.DollarNames}{ReducedExperiment}(x, pattern = "")
}
\arguments{
\item{x}{The \link[ReducedExperiment]{ReducedExperiment} object.}

\item{pattern}{Search pattern.}
}
\value{
The names of the matching columns of \code{colData}.
}
\description{
Command line completion for \code{$}.
This function is not intended to be used directly by users but provides
auto-completion capabilities.
Autocompletes based on column data names (i.e., the column names of the
\code{colData}).
}
\seealso{
\code{\link[utils:rcompgen]{utils::.DollarNames()}}
}
