% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{.adjust_by_term}
\alias{.adjust_by_term}
\title{Adjusts the p-values of model results for multiple testing on a per-term basis}
\usage{
.adjust_by_term(res, method = "BH")
}
\description{
Adjusts the p-values of model results for multiple testing on a per-term basis
}
