% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{.calculate_eigengenes}
\alias{.calculate_eigengenes}
\title{Calculates eigengenes for new data based on module assignments}
\usage{
.calculate_eigengenes(
  newdata,
  module_names,
  module_assignments,
  realign = TRUE
)
}
\description{
Calculates eigengenes for new data based on module assignments
}
