% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{.createRandomisedModularExperiment}
\alias{.createRandomisedModularExperiment}
\title{i (features), j (samples), k components)}
\usage{
.createRandomisedModularExperiment(i, j, k)
}
\description{
i (features), j (samples), k components)
}
