% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{.createRandomisedReducedExperiment}
\alias{.createRandomisedReducedExperiment}
\title{Create a ReducedExperiment with random data}
\usage{
.createRandomisedReducedExperiment(i, j, k, seed = 1)
}
\arguments{
\item{i}{Number of features}

\item{j}{Number of samples}

\item{k}{Number of dimensionally-reduced components}

\item{seed}{Random seed that will be set before call to rnorm}
}
\description{
Creates a \link[ReducedExperiment]{ReducedExperiment}
with i (features), j (samples), k components)
}
