% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R
\name{.DollarNames.FactorisedExperiment}
\alias{.DollarNames.FactorisedExperiment}
\title{Required for dollarsign autocomplete of colData columns}
\usage{
\method{.DollarNames}{FactorisedExperiment}(x, pattern = "")
}
\description{
Required for dollarsign autocomplete of colData columns
}
