% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R
\name{.DollarNames.ModularExperiment}
\alias{.DollarNames.ModularExperiment}
\title{Required for dollarsign autocomplete of colData columns}
\usage{
\method{.DollarNames}{ModularExperiment}(x, pattern = "")
}
\description{
Required for dollarsign autocomplete of colData columns
}
