% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ReducedExperiment.R
\name{.DollarNames.ReducedExperiment}
\alias{.DollarNames.ReducedExperiment}
\title{Required for dollarsign autocomplete of colData columns}
\usage{
\method{.DollarNames}{ReducedExperiment}(x, pattern = "")
}
\description{
Required for dollarsign autocomplete of colData columns
}
