% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{.format_enrich_res}
\alias{.format_enrich_res}
\title{Reformats enrichment result}
\usage{
.format_enrich_res(enrich_res_single, adj_method, p_cutoff, min_genes = NULL)
}
\description{
Reformats \code{clusterProfiler} enrichment results into a consistent format.
}
