% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{.get_airway_data}
\alias{.get_airway_data}
\title{Gets data from the airway package}
\usage{
.get_airway_data(n_features = NULL)
}
\arguments{
\item{n_features}{If not NULL, the number of features (genes) to be randomly
selected from the original airway data. Can provide a smaller dataset for
testing}
}
\description{
Gets data from the airway package and returns it as a SummarizedExperiment
with counts ("counts") and log-transformed counts ("normal") in the assay
slots. Non-expressed genes are removed.
}
