% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{.get_t2g}
\alias{.get_t2g}
\title{Gets a TERM2GENE data object for clusterProfiler}
\usage{
.get_t2g(database, TERM2GENE)
}
\description{
Gets a correctly formatted TERM2GENE object for use by the \code{clusterProfiler}
package for enrichment analysis.
}
