% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{.makeRandomData}
\alias{.makeRandomData}
\title{Turn rnorm into a matrix}
\usage{
.makeRandomData(r, c, rname, cname, seed = 1)
}
\arguments{
\item{r}{Number of rows}

\item{c}{Number of columns}

\item{rname}{Name of rows (will be appended to the row number)}

\item{cname}{Name of columns (will be appended to the column number)}

\item{seed}{Random seed that will be set before call to rnorm}
}
\description{
Used to generate a matrix of random numbers with dimensions r (rows) * c
(columns)
}
