% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{.project_eigengenes}
\alias{.project_eigengenes}
\title{Calculates eigengenes for new data based on the original PCA rotation matrix}
\usage{
.project_eigengenes(
  newdata,
  module_names,
  module_assignments,
  lod,
  min_module_genes
)
}
\description{
Calculates eigengenes for new data based on the original PCA rotation matrix
}
