% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{.rename_results_table}
\alias{.rename_results_table}
\title{Renames linear model result table column names}
\usage{
.rename_results_table(cnames)
}
\description{
Renames linear model result table column names
}
