% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{.reorient_factors}
\alias{.reorient_factors}
\title{Ensure factors have positive skew}
\usage{
.reorient_factors(S)
}
\description{
Reorientates factors based on their skew. Generally ensures that the most
aligned genes are positively associated with the factor.
}
