% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{.run_linear_model}
\alias{.run_linear_model}
\title{Helper function to run a linear model}
\usage{
.run_linear_model(X, pheno, formula, method = "lm", type = "II", ...)
}
\description{
Helper function to run a linear model
}
