% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{.se_to_fe}
\alias{.se_to_fe}
\title{Creates a FactorisedExperiment from a SummarizedExperiment}
\usage{
.se_to_fe(se, reduced, loadings, stability, center_X, scale_X)
}
\arguments{
\item{se}{A \link{SummarizedExperiment}\link{SummarizedExperiment} object.}

\item{reduced}{Data to be passed to the \code{reduced} slot.}

\item{loadings}{Data to be passed to the \code{loadings} slot.}

\item{stability}{Data to be passed to the \code{stability} slot.}

\item{center_X}{Data to be passed to the \code{center_X} slot.}

\item{scale_X}{Data to be passed to the \code{scale_X} slot.}
}
\description{
Helper function for transforming a
\link{ReducedExperiment}\link{FactorisedExperiment} into a
\link{SummarizedExperiment}\link{SummarizedExperiment}
}
