% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{.se_to_me}
\alias{.se_to_me}
\title{Creates a ModularExperiment from a SummarizedExperiment}
\usage{
.se_to_me(
  se,
  reduced,
  loadings,
  assignments,
  center_X,
  scale_X,
  dendrogram = NULL,
  threshold = NULL
)
}
\arguments{
\item{se}{A \link{SummarizedExperiment}\link{SummarizedExperiment} object.}

\item{reduced}{Data to be passed to the \code{reduced} slot.}

\item{loadings}{Data to be passed to the \code{loadings} slot.}

\item{assignments}{Data to be passed to the \code{assignments} slot.}

\item{center_X}{Data to be passed to the \code{center_X} slot.}

\item{scale_X}{Data to be passed to the \code{scale_X} slot.}

\item{dendrogram}{Data to be passed to the \code{dendrogram} slot.}

\item{threshold}{Data to be passed to the \code{threshold} slot.}
}
\description{
Helper function for transforming a
\link{ReducedExperiment}\link{ModularExperiment} into a
\link{SummarizedExperiment}\link{SummarizedExperiment}
}
