% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples.R
\name{.single_lmer}
\alias{.single_lmer}
\title{Helper function to run a linear mixed model}
\usage{
.single_lmer(data, formula_string, REML = TRUE, ...)
}
\description{
Helper function to run a linear mixed model
}
