% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{.stability_ica}
\alias{.stability_ica}
\title{Stability ICA method}
\usage{
.stability_ica(
  X,
  nc = nc,
  resample = resample,
  method = method,
  n_runs = 30,
  ...
)
}
\description{
Stability ICA method
}
