% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R,
%   R/methods-ModularExperiment.R
\name{runEnrich,FactorisedExperiment-method}
\alias{runEnrich,FactorisedExperiment-method}
\alias{runEnrich,ModularExperiment-method}
\title{Functional enrichment analyses for dimensionally-reduced data}
\usage{
\S4method{runEnrich}{FactorisedExperiment}(
  object,
  method = "gsea",
  feature_id_col = "rownames",
  center_loadings = FALSE,
  abs_loadings = FALSE,
  loading_threshold = 0.5,
  proportional_threshold = 0.01,
  as_dataframe = FALSE,
  ...
)

\S4method{runEnrich}{ModularExperiment}(
  object,
  method = "overrepresentation",
  feature_id_col = "rownames",
  as_dataframe = FALSE,
  ...
)
}
\arguments{
\item{object}{\link[ReducedExperiment]{FactorisedExperiment}  or
\link[ReducedExperiment]{ModularExperiment} object.}

\item{method}{The method to use for identifying enriched pathways. One of
"overrepresentation" or "gsea". The overrepresentation method calls
\link[clusterProfiler]{enricher} whereas the gsea method calls
\link[clusterProfiler]{GSEA}. Note that GSEA is not available for modules.}

\item{feature_id_col}{The column in \code{rowData(object)} that will be used as a
feature ID. Setting this to "rownames" (default) instead uses
\code{rownames(object)}.}

\item{center_loadings}{Factors only: If \code{TRUE}, loadings will be centered
column-wise to have a mean of 0.}

\item{abs_loadings}{Factors only: If \code{TRUE}, the absolute values of the
loadings will be used for enrichment analysis. If \code{FALSE}, the signed
loadings will be used for GSEA enrichment. Note that, regardless of the
value of this term, the process used to select genes for overrepresentation
analysis will be based on absolute loadings.}

\item{loading_threshold}{Factors only: See
\link[ReducedExperiment]{getAlignedGenes}. Only relevant for overresentation
analysis.}

\item{proportional_threshold}{Factors only: See
\link[ReducedExperiment]{getAlignedGenes}. Only relevant for overresentation
analysis.}

\item{as_dataframe}{If \code{TRUE}, the results will be returned as a data.frame.
Otherwise, the results will be returned as a list of objects created by either
\link[clusterProfiler]{enricher}, in the case of overrepresentation
analysis, or \link[clusterProfiler]{GSEA}, in the case of GSEA.}
}
\description{
Functional enrichment analyses for dimensionally-reduced data
}
\details{
When running module analysis, the overrepresentation method identifies
pathways that are overrepresented in each module.

For factor analysis, the overrepresentation method first identifies the genes
most highly aligned with each factor
(using \link[ReducedExperiment]{getAlignedGenes}), then uses
the resulting gene lists to perform overrepresentation analysis. The GSEA
method instead uses the entire set of factor loadings, and identifies pathways
that are overrepresented in the tails of this distribution.
}
