% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{estimate_factors}
\alias{estimate_factors}
\title{Apply dimensionality reduction using ICA}
\usage{
estimate_factors(
  X,
  nc,
  center_X = TRUE,
  scale_X = FALSE,
  assay_name = "normal",
  ...
)
}
\description{
Apply dimensionality reduction using ICA
}
