% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{estimate_stability}
\alias{estimate_stability}
\title{Estimate stability of factors as a function of the number of components}
\usage{
estimate_stability(
  X,
  min_components = 10,
  max_components = 60,
  by = 2,
  n_runs = 30,
  resample = FALSE,
  mean_stability_threshold = NULL,
  center_X = TRUE,
  scale_X = FALSE,
  BPPARAM = BiocParallel::SerialParam(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{X}{Either a \link{SummarizedExperiment}\link{SummarizedExperiment} object
or a matrix containing data to be subject to ICA. \code{X} should have rows as
features and columns as samples. If a
\link{SummarizedExperiment}\link{SummarizedExperiment} is passed, it is assumed
that there is an assay named "normal" that contains the input data.}

\item{min_components}{The minimum number of components to estimate the
stability for.}

\item{max_components}{The maximum number of components to estimate the
stability for.}

\item{by}{The number by which to increment the numbers of components
tested.}

\item{n_runs}{The number of times to run ICA. Ignored if \code{use_stability} is
FALSE. See \code{details} for further information.}

\item{resample}{If TRUE, a boostrap approach is used to estimating factors.
Else, random initialisation of ICA is employed. Ignored if \code{use_stability}
is FALSE. See \code{details} for further information.}

\item{mean_stability_threshold}{The function will return the minimal
number of components that exceed this stability threshold.}

\item{center_X}{If TRUE, X is centered (i.e., features / rows are transformed
to have a mean of 0) prior to ICA. Generally recommended.}

\item{scale_X}{If TRUE, X is scaled (i.e., features / rows are transformed
to have a standard deviation of 1) before ICA.}

\item{BPPARAM}{A class containing parameters for parallel evaluation. Uses
\link[BiocParallel]{SerialParam} by default, running only a single
ICA computation at a time. Ignored if \code{use_stability}
is FALSE. See \code{details} for further information.}

\item{verbose}{If TRUE, shows a progress bar that updates for each
number of components tested. Note that the time taken may not be linear,
because the time taken to run ICA generally increases with the number
of components.}

\item{...}{Additional arguments to be passed to
\link{ReducedExperiment}\link{run_ica}.}
}
\value{
Returns a list containing: i) a list of stabilities as a function
of the number of components; and ii) the selected number of components based
on the \code{mean_stability_threshold}.
}
\description{
Estimates the stability of factors over a range of component numbers to
aid in the identification of the optimal factor number.
}
\details{
Runs the stability-based ICA algorithm
(see \link[ReducedExperiment]{run_ica}) for a range of component numbers.
Estimates stability for each one, allowing for selection of the optimal
number of components to be used for ICA. The results of this function
can be plotted by \link[ReducedExperiment]{plot_stability}.

This algorithm is similar to the Maximally Stable Transcriptome
Dimension (MSTD) approach
(https://bmcgenomics.biomedcentral.com/articles/10.1186/s12864-017-4112-9).

The function automatically selects a number of components based on
\code{mean_stability_threshold}. Ideally, this should not be trusted blindly,
but instead a choice can be made based on visualisation of the
stabilities, as created by \link[ReducedExperiment]{plot_stability}. The
MSTD paper provides additional context and advice on choosing the number
of components based on these data.
}
