% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R
\name{loadings,FactorisedExperiment-method}
\alias{loadings,FactorisedExperiment-method}
\alias{loadings<-,FactorisedExperiment-method}
\title{Get and set factor loadings}
\usage{
\S4method{loadings}{FactorisedExperiment}(
  object,
  scale_loadings = FALSE,
  center_loadings = FALSE,
  abs_loadings = FALSE
)

\S4method{loadings}{FactorisedExperiment}(object) <- value
}
\arguments{
\item{object}{\link[ReducedExperiment]{FactorisedExperiment} object.}

\item{scale_loadings}{If \code{TRUE}, loadings will be scaled column-wise to have a
standard deviation of 0.}

\item{center_loadings}{If \code{TRUE}, loadings will be centered column-wise to have a mean
of 0.}

\item{abs_loadings}{If \code{TRUE}, the absolute values of the loadings will be
returned.}
}
\description{
Retrieves the loadings matrix, with features as rows and reduced
components as columns.
}
\details{
The loadings can be modified with \verb{<-}.
}
