% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-FactorisedExperiment.R
\docType{class}
\name{FactorisedExperiment-class}
\alias{FactorisedExperiment-class}
\alias{.FactorisedExperiment}
\alias{FactorisedExperiment}
\title{FactorisedExperiment: A container for the results of factor analysis}
\usage{
FactorisedExperiment(
  reduced = new("matrix"),
  scale = TRUE,
  center = TRUE,
  loadings = new("matrix"),
  stability = NULL,
  ...
)
}
\arguments{
\item{reduced}{A data matrix, produced by factor analysis, with rows
representing samples and columns representing factors.}

\item{scale}{Either a boolean, representing whether or not the original data
has been scaled to unit variance, or a numeric vector indicating the
standard deviations of the original features (as produced by
\link[base]{scale}.)}

\item{center}{Either a boolean, representing whether or not the original data
has been centered to have a mean of 0, or a numeric vector indicating the
means of the original features (as produced by
\link[base]{scale}.)}

\item{loadings}{A data matrix, produced by factor analysis, with rows
representing features and columns representing factors.}

\item{stability}{A vector containing some measure of stability or variance
explained for each factor. If factor analysis was performed using
\link[ReducedExperiment]{estimate_factors} and \code{use_stability = True}, this
slot will indicate the stability of the factors across multiple runs of ICA.}

\item{...}{Additional arguments to be passed to
\link[ReducedExperiment]{ReducedExperiment}.}
}
\description{
A container inheriting from \link[ReducedExperiment]{ReducedExperiment}, that
contains one or more data matrices, to which factor analysis has been applied
to identify a reduced set of features.
}
