% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R
\name{FactorisedExperiment}
\alias{FactorisedExperiment}
\title{FactorisedExperiment}
\usage{
FactorisedExperiment(
  loadings = new("matrix"),
  stability = new("numeric"),
  scale = FALSE,
  center = FALSE,
  ...
)
}
\description{
A class to represent the results of factor analysis
}
