% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R,
%   R/methods-ModularExperiment.R, R/methods-ReducedExperiment.R
\name{names<-,FactorisedExperiment-method}
\alias{names<-,FactorisedExperiment-method}
\alias{featureNames<-,FactorisedExperiment-method}
\alias{rownames<-,FactorisedExperiment-method}
\alias{names<-,ModularExperiment-method}
\alias{featureNames<-,ModularExperiment-method}
\alias{rownames<-,ModularExperiment-method}
\alias{featureNames,ReducedExperiment-method}
\alias{names<-,ReducedExperiment-method}
\alias{rownames<-,ReducedExperiment-method}
\alias{featureNames<-,ReducedExperiment-method}
\title{Get feature names}
\usage{
\S4method{names}{FactorisedExperiment}(x) <- value

\S4method{featureNames}{FactorisedExperiment}(object) <- value

\S4method{rownames}{FactorisedExperiment}(x) <- value

\S4method{names}{ModularExperiment}(x) <- value

\S4method{featureNames}{ModularExperiment}(object) <- value

\S4method{rownames}{ModularExperiment}(x) <- value

\S4method{featureNames}{ReducedExperiment}(object)

\S4method{names}{ReducedExperiment}(x) <- value

\S4method{rownames}{ReducedExperiment}(x) <- value

\S4method{featureNames}{ReducedExperiment}(object) <- value
}
\arguments{
\item{object}{\link[ReducedExperiment]{ReducedExperiment} object.}
}
\description{
Retrieves feature names (usually genes).
}
\details{
Feature names can be modified with \verb{<-}.
}
