% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R
\name{getCentrality,ModularExperiment-method}
\alias{getCentrality,ModularExperiment-method}
\title{Get correlation of features with module eigengenes}
\usage{
\S4method{getCentrality}{ModularExperiment}(object, assay_name = "normal", feature_id_col = "rownames")
}
\arguments{
\item{object}{\link[ReducedExperiment]{ModularExperiment} object.}

\item{assay_name}{The name of the assay to be used for calculation of
module centrality.}

\item{feature_id_col}{The column in \code{rowData(object)} that will be used as a
feature ID. Setting this to "rownames" (default) instead uses
\code{rownames(object)}.

Results indicate correlation (r) and squared correlation (rsq) with the module eigengene. Feature ranks
within each module are also returned based on these statistics.}
}
\description{
Get correlation of features with module eigengenes
}
