% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ReducedExperiment.R
\name{getGeneIDs,ReducedExperiment-method}
\alias{getGeneIDs,ReducedExperiment-method}
\title{Gets alternative gene annotations from biomaRt}
\usage{
\S4method{getGeneIDs}{ReducedExperiment}(
  object,
  gene_id_col = "rownames",
  gene_id_type = "ensembl_gene_id",
  ids_to_get = c("hgnc_symbol", "entrezgene_id"),
  dataset = "hsapiens_gene_ensembl",
  mart = NULL
)
}
\arguments{
\item{object}{\link[ReducedExperiment]{ReducedExperiment} object.}

\item{gene_id_col}{The column in \code{rowData(object)} that will be used to
query biomaRt. Setting this to "rownames" instead uses \code{rownames(object)}
for matching.}

\item{gene_id_type}{The type of attribute to be used to query with biomaRt.
See the \code{filters} argument of \link[biomaRt]{getBM}.}

\item{ids_to_get}{The type of attribute to get from biomaRt.
See the \code{attributes} argument of \link[biomaRt]{getBM}.}

\item{dataset}{The Ensembl dataset to retrieve. See the \code{dataset} argument
of \link[biomaRt]{useEnsembl}. If \code{mart} is not NULL, this argument is
ignored.}

\item{mart}{An optional mart object to use. See the \code{mart} argument of
\link[biomaRt]{getBM}.

@returns Returns the original object, with additional variables added to
the \code{rowData} slot.}
}
\description{
Uses \link[biomaRt]{getBM} to get alternative gene IDs for
\link[ReducedExperiment]{ReducedExperiment} objects. The new annotations
are added as columns to the input object's \code{rowData}
}
