% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R,
%   R/methods-ReducedExperiment.R
\name{nModules,ModularExperiment-method}
\alias{nModules,ModularExperiment-method}
\alias{nComponents,ReducedExperiment-method}
\alias{nSamples,ReducedExperiment-method}
\alias{nFeatures,ReducedExperiment-method}
\title{Prints individual lengths of samples, components and features}
\usage{
\S4method{nModules}{ModularExperiment}(object)

\S4method{nComponents}{ReducedExperiment}(object)

\S4method{nSamples}{ReducedExperiment}(object)

\S4method{nFeatures}{ReducedExperiment}(object)
}
\arguments{
\item{object}{\link[ReducedExperiment]{ReducedExperiment} object.}
}
\value{
The number of samples (\code{nSamples}), features (\code{nFeatures})
and dimensionally-reduced components (\code{nComponents}) are returned.
}
\description{
Prints individual lengths of samples, components and features
}
