% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R
\name{ModularExperiment}
\alias{ModularExperiment}
\title{ModularExperiment: A container for the results of module analysis}
\usage{
ModularExperiment(
  loadings = numeric(),
  assignments = character(),
  dendrogram = NULL,
  threshold = NULL,
  ...
)
}
\arguments{
\item{loadings}{A numeric vector representing the loadings used to generate
module expression profiles. Typically, these values are obtained from the
rotation matrix produced by \link[stats]{prcomp}, which is used to identify
the first principal component of each module. The vector names represent
features.}

\item{assignments}{A vector of features, named according to the module to which the
feature belongs.}

\item{dendrogram}{Either NULL, or the dendrogram used to identify modules from the
original data.}

\item{threshold}{Either NULL, or a matrix produced by
\link[WGCNA]{pickSoftThreshold} indicating the parameters used for network
construction.

A class to represent the results of module analysis}

\item{reduced}{A data matrix, produced by module analysis, with rows
representing samples and columns representing module expression profiles.
Typically, this matrix contains "eigengenes" produced by the Weighted Gene
Correlation Network Analysis Approach, as is applied by
\link[ReducedExperiment]{identify_modules}.}

\item{scale}{Either a boolean, representing whether or not the original data
has been scaled to unit variance, or a numeric vector indicating the
standard deviations of the original features (as produced by
\link[base]{scale}.)}

\item{center}{Either a boolean, representing whether or not the original data
has been centered to have a mean of 0, or a numeric vector indicating the
means of the original features (as produced by
\link[base]{scale}.)}
}
\description{
A container inheriting from \link[ReducedExperiment]{ReducedExperiment}, that
contains one or more data matrices, to which module analysis has been applied
to identify a reduced set of features.
}
