% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R
\name{loadings,ModularExperiment-method}
\alias{loadings,ModularExperiment-method}
\alias{loadings<-,ModularExperiment-method}
\title{Get and set modules loadings}
\usage{
\S4method{loadings}{ModularExperiment}(
  object,
  scale_loadings = FALSE,
  center_loadings = FALSE,
  abs_loadings = FALSE
)

\S4method{loadings}{ModularExperiment}(object) <- value
}
\arguments{
\item{object}{\link[ReducedExperiment]{ModularExperiment} object.}

\item{scale_loadings}{If \code{TRUE}, loadings will be scaled column-wise to have a
standard deviation of 0.}

\item{center_loadings}{If \code{TRUE}, loadings will be centered column-wise to have a mean
of 0.}

\item{abs_loadings}{If \code{TRUE}, the absolute values of the loadings will be
returned.}
}
\description{
Retrieves the module loadings as a vector, with a value for each feature
(usually genes).
}
\details{
The loadings can be modified with \verb{<-}.

When available, the module loadings provide the values of the rotation matrix
(usually generated by \link[stats]{prcomp}) used to calculate the
sample-level module vectors available in the \code{reduced} slot. Normally, these
loadings are calculated for each module separately, so their values are
not comparable across modules.
}
