% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{module_preservation}
\alias{module_preservation}
\title{Get module preservation statistics}
\usage{
module_preservation(
  reference_dataset,
  test_dataset,
  reference_assay_name = "normal",
  test_assay_name = "normal",
  module_assignments = NULL,
  networkType = "signed",
  corFnc = "cor",
  ...
)
}
\arguments{
\item{reference_dataset}{The dataset that was used to define the modules.
Must be a \code{data.frame} or \code{matrix} with features as rows and samples as
columns, or a \link[ReducedExperiment]{ModularExperiment} or
\link[SummarizedExperiment]{SummarizedExperiment} object.}

\item{test_dataset}{The dataset that will be used to test for module
preservation. Must be a \code{data.frame} or \code{matrix} with features as rows and
samples as columns, or a link[SummarizedExperiment]{SummarizedExperiment}
object. The features of \code{test_dataset} should be the same as
\code{reference_dataset}.}

\item{reference_assay_name}{If the reference dataset is a
\link[ReducedExperiment]{ModularExperiment} or
\link[SummarizedExperiment]{SummarizedExperiment} object, this argument
specifies which assay slot was used to define the modules.}

\item{test_assay_name}{If the reference dataset is a
\link[ReducedExperiment]{ModularExperiment} or
\link[SummarizedExperiment]{SummarizedExperiment} object, this argument
specifies which assay slot is to be used in preservation tests.}

\item{module_assignments}{If the reference dataset is not a
\link[ReducedExperiment]{ModularExperiment} object, this argument is
necessary to specify the module assignments.}
}
\value{
A \code{data.frame} containing preservation statistics, as described
by \link[WGCNA]{modulePreservation}.
}
\description{
Tests whether a set of modules defined in the reference dataset are
preserved in the test dataset.
}
