% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicsDimReduce.R
\name{omicsDimReduce}
\alias{omicsDimReduce}
\title{omicsDimReduce}
\description{
A toolbox for reducing the dimensionality of expression data and
analysing the decomposed data. Provides utilities for independent component
analysis (ICA) and weighted gene correlation network analysis (WGCNA).
}
\seealso{
\code{\link{reduce_data}}
}
\author{
Jack Gisby
}
