% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{plot_common_features}
\alias{plot_common_features}
\title{Heatmap comparing commonality across factors}
\usage{
plot_common_features(
  common_features,
  filename = NA,
  color = (grDevices::colorRampPalette(RColorBrewer::brewer.pal(n = 7, name =
    "YlOrRd")))(100)
)
}
\arguments{
\item{common_features}{The output of
\link[SummarizedExperiment]{get_common_features}.}

\item{filename}{The path at which to save the plot.}

\item{color}{The colour palette to be used in the heatmap.}
}
\value{
An object generated by \link[pheatmap]{pheatmap}.
}
\description{
Heatmap comparing commonality across factors
}
