% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{plot_module_preservation}
\alias{plot_module_preservation}
\title{Plot module preservation statistics}
\usage{
plot_module_preservation(module_preservation_results)
}
\arguments{
\item{module_preservation_results}{The output of
\link[ReducedExperiment]{module_preservation}}
}
\description{
Plot module preservation statistics
}
