% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{plot_stability}
\alias{plot_stability}
\title{Plot component stability as a function of the number of components}
\usage{
plot_stability(
  stability,
  plot_path,
  stability_threshold = NULL,
  mean_stability_threshold = NULL,
  height = 4,
  width = 10,
  ...
)
}
\description{
Plot component stability as a function of the number of components
}
\references{
MSTD
}
