% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorSet.R
\name{project_data,FactorisedExperiment,matrix-method}
\alias{project_data,FactorisedExperiment,matrix-method}
\title{Project data}
\usage{
\S4method{project_data}{FactorisedExperiment,matrix}(x, newdata, new_data_is_transformed = FALSE)
}
\description{
Project data
}
