% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R
\name{projectData,FactorisedExperiment,matrix-method}
\alias{projectData,FactorisedExperiment,matrix-method}
\title{Project data}
\usage{
\S4method{projectData}{FactorisedExperiment,matrix}(x, newdata)
}
\description{
Project data
}
