% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data.R
\name{reduce_data}
\alias{reduce_data}
\title{Apply dimensionality reduction}
\usage{
reduce_data(
  X,
  method = c("ICA", "WGCNA"),
  phenoData = Biobase::annotatedDataFrameFrom(X, byrow = FALSE),
  featureData = Biobase::annotatedDataFrameFrom(X, byrow = TRUE),
  center_X = (method == "ICA"),
  scale_X = FALSE,
  experimentData = MIAME(),
  annotation = character(),
  protocolData = annotatedDataFrameFrom(X, byrow = FALSE),
  ...
)
}
\description{
Apply dimensionality reduction
}
