% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ReducedExperiment.R
\name{reduced,ReducedExperiment-method}
\alias{reduced,ReducedExperiment-method}
\alias{reduced<-,ReducedExperiment-method}
\title{Get and set reduced data}
\usage{
\S4method{reduced}{ReducedExperiment}(object, scale_reduced = FALSE, center_reduced = FALSE)

\S4method{reduced}{ReducedExperiment}(object) <- value
}
\arguments{
\item{object}{\link[ReducedExperiment]{ReducedExperiment} object.}

\item{scale_reduced}{If \code{TRUE}, data will be scaled column-wise to have a
standard deviation of 0.}

\item{center_reduced}{If \code{TRUE}, data will be center4ed column-wise to have a mean
of 0.}
}
\description{
Retrieves the reduced data matrix, with samples as rows and reduced
components as columns.
}
\details{
The reduced data can be modified with \verb{<-}.
}
