% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{reduced_oa}
\alias{reduced_oa}
\title{Overrepresentation analysis}
\usage{
reduced_oa(
  component_features,
  database = "msigdb_c2_cp",
  TERM2GENE = NULL,
  p_cutoff = 1,
  adj_method = "BH",
  min_genes = 3,
  universe = NULL,
  ...
)
}
\description{
Performs overrepresentation analysis given a set of components (factors or
modules) that are each linked to a list of genes.
}
