% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{reduced_overrepresentation}
\alias{reduced_overrepresentation}
\title{Overrepresentation analysis}
\usage{
reduced_overrepresentation(
  component_features,
  database = "msigdb_c2_cp",
  TERM2GENE = NULL,
  p_cutoff = 1,
  adj_p_cutoff = 1,
  adj_method = "BH",
  min_genes = 3,
  universe = NULL,
  ...
)
}
\description{
Overrepresentation analysis
}
