% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReducedExperiment.R
\name{ReducedExperiment-package}
\alias{ReducedExperiment-package}
\title{ReducedExperiment}
\description{
A toolbox for reducing the dimensionality of expression data and
analysing the decomposed data. Provides utilities for independent component
analysis (ICA) and weighted gene correlation network analysis (WGCNA).
}
\author{
Jack Gisby
}
