% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{run_ica}
\alias{run_ica}
\title{Run ICA for a data matrix}
\usage{
run_ica(
  X,
  nc,
  method = "imax",
  center_X = TRUE,
  scale_X = FALSE,
  reorient_skewed = TRUE,
  seed = 1,
  ...
)
}
\description{
Run ICA for a data matrix
}
