% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{run_wgcna}
\alias{run_wgcna}
\title{Run WGCNA for a data matrix}
\usage{
run_wgcna(
  X,
  powers = 1:30,
  min_r_squared = 0.85,
  max_mean_connectivity = 100,
  corType = "pearson",
  networkType = "signed",
  module_labels = "numbers",
  maxBlockSize = 30000,
  seed = 1,
  verbose = 0,
  return_full_output = FALSE,
  scale_reduced = TRUE,
  ...
)
}
\description{
Run WGCNA for a data matrix
}
