% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ReducedExperiment.R
\name{sampleNames,ReducedExperiment-method}
\alias{sampleNames,ReducedExperiment-method}
\alias{sampleNames<-,ReducedExperiment-method}
\title{Get sample names}
\usage{
\S4method{sampleNames}{ReducedExperiment}(object)

\S4method{sampleNames}{ReducedExperiment}(object) <- value
}
\arguments{
\item{object}{\link[ReducedExperiment]{ReducedExperiment} object.}
}
\description{
Retrieves sample names.
}
\details{
Sample names can be modified with \verb{<-}.
}
