% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ReducedExperiment.R
\name{show,ReducedExperiment-method}
\alias{show,ReducedExperiment-method}
\title{Prints a summary of a ReducedExperiment object}
\usage{
\S4method{show}{ReducedExperiment}(object)
}
\arguments{
\item{object}{\link[ReducedExperiment]{ReducedExperiment} object.}
}
\description{
Prints a summary of a ReducedExperiment object
}
