% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R
\name{stability,FactorisedExperiment-method}
\alias{stability,FactorisedExperiment-method}
\alias{stability<-,FactorisedExperiment-method}
\title{Gets the stability values for factors}
\usage{
\S4method{stability}{FactorisedExperiment}(object)

\S4method{stability}{FactorisedExperiment}(object) <- value
}
\arguments{
\item{object}{\link[ReducedExperiment]{FactorisedExperiment} object.}
}
\description{
The stability values can be modified with \verb{<-}.
}
