% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{.createRandomisedFactorisedExperiment}
\alias{.createRandomisedFactorisedExperiment}
\title{i (features), j (samples), k components)}
\usage{
.createRandomisedFactorisedExperiment(i, j, k)
}
\description{
i (features), j (samples), k components)
}
