% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{.createRandomisedFactorisedExperiment}
\alias{.createRandomisedFactorisedExperiment}
\title{Create a FactorisedExperiment with random data}
\usage{
.createRandomisedFactorisedExperiment(i, j, k, seed = 1)
}
\arguments{
\item{i}{Number of features}

\item{j}{Number of samples}

\item{k}{Number of factors}

\item{seed}{Random seed that will be set before call to rnorm}
}
\description{
Creates a \link[ReducedExperiment]{FactorisedExperiment}
with i (features), j (samples), k components)
}
