% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{.createRandomisedReducedExperiment}
\alias{.createRandomisedReducedExperiment}
\title{i (features), j (samples), k components)}
\usage{
.createRandomisedReducedExperiment(i, j, k)
}
\description{
i (features), j (samples), k components)
}
