% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorSet.R
\name{FactorisedExperiment}
\alias{FactorisedExperiment}
\title{FactorisedExperiment}
\usage{
FactorisedExperiment(loadings = new("matrix"), varexp = new("numeric"), ...)
}
\description{
A class to represent the results of factor analysis
}
