% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R,
%   R/methods-ModularExperiment.R, R/methods-ReducedExperiment.R
\name{names<-,FactorisedExperiment-method}
\alias{names<-,FactorisedExperiment-method}
\alias{featureNames<-,FactorisedExperiment-method}
\alias{rownames<-,FactorisedExperiment-method}
\alias{names<-,ModularExperiment-method}
\alias{featureNames<-,ModularExperiment-method}
\alias{rownames<-,ModularExperiment-method}
\alias{featureNames}
\alias{featureNames<-}
\alias{featureNames,ReducedExperiment-method}
\alias{names<-,ReducedExperiment-method}
\alias{rownames<-,ReducedExperiment-method}
\alias{featureNames<-,ReducedExperiment-method}
\title{Get feature names}
\usage{
\S4method{names}{FactorisedExperiment}(x) <- value

\S4method{featureNames}{FactorisedExperiment}(x) <- value

\S4method{rownames}{FactorisedExperiment}(x) <- value

\S4method{names}{ModularExperiment}(x) <- value

\S4method{featureNames}{ModularExperiment}(x) <- value

\S4method{rownames}{ModularExperiment}(x) <- value

\S4method{featureNames}{ReducedExperiment}(x)

\S4method{names}{ReducedExperiment}(x) <- value

\S4method{rownames}{ReducedExperiment}(x) <- value

\S4method{featureNames}{ReducedExperiment}(x) <- value
}
\arguments{
\item{x}{\link[ReducedExperiment]{ReducedExperiment} object.}

\item{value}{New value to replace existing names.}
}
\description{
Retrieves feature names (rownames, usually genes).
}
