% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{get_msigdb_t2g}
\alias{get_msigdb_t2g}
\title{Get TERM2GENE dataframe from MSigDB}
\usage{
get_msigdb_t2g(
  species = "Homo sapiens",
  category = "C2",
  subcategory = NULL,
  subcategory_to_remove = "CGP",
  gene_id = "ensembl_gene"
)
}
\description{
Get TERM2GENE dataframe from MSigDB
}
