% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{identify_modules}
\alias{identify_modules}
\title{Apply dimensionality reduction using WGCNA}
\usage{
identify_modules(
  X,
  center_X = TRUE,
  scale_X = TRUE,
  assay_name = "normal",
  ...
)
}
\arguments{
\item{X}{Either a \link{SummarizedExperiment}\link{SummarizedExperiment} object
or a matrix containing data to be subject to WGCNA. \code{X} should have rows as
features and columns as samples.}

\item{center_X}{If TRUE, X is centered (i.e., features / rows are transformed
to have a mean of 0) prior to WGCNA.}

\item{scale_X}{If TRUE, X is scaled (i.e., features / rows are transformed
to have a standard deviation of 1) before WGCNA.}

\item{assay_name}{If \code{X} is a
\link{SummarizedExperiment}\link{SummarizedExperiment}, then this should be the
name of the assay to be subject to WGCNA.}

\item{...}{Additional arguments to be passed to
\link{ReducedExperiment}\link{run_wgcna}.}
}
\value{
A \link{ReducedExperiment}\link{ModularExperiment} is returned
containing the input data (i.e., the original data matrix in addition to
other slots if a \link{SummarizedExperiment}\link{SummarizedExperiment} was used
as input). Additionally contains the results of module analysis, stored in
the \code{reduced} and \code{assignments} slots. The \code{center_X}, \code{scale_X},
\code{loadings}, \code{threshold} and \code{dendrogram} slots may also be filled depending
on the arguments given to \code{identify_modules}.
}
\description{
Performs Weighted gene correlation network analysis (WGCNA).
Calls \link{ReducedExperiment}\link{run_wgcna} to perform the analysis.
}
\examples{
# Get the airway data as a SummarizedExperiment (with a subset of features)
set.seed(2)
airway_se <- ReducedExperiment:::.get_airway_data(n_features = 500)

# Identify modules in the airway data
WGCNA::disableWGCNAThreads()
me_1 <- identify_modules(airway_se, verbose = 0, powers = 5)
me_1

}
\seealso{
\code{\link[=run_wgcna]{run_wgcna()}},
\code{\link[WGCNA:blockwiseModules]{WGCNA::blockwiseModules()}},
\code{\link[WGCNA:pickSoftThreshold]{WGCNA::pickSoftThreshold()}}
}
\author{
Jack Gisby
}
