% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R
\name{ModularExperiment}
\alias{ModularExperiment}
\title{ModularExperiment}
\usage{
ModularExperiment(
  assignments = character(),
  dendrogram = NULL,
  threshold = NULL,
  ...
)
}
\description{
A class to represent the results of module analysis
}
