% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModuleSet.R
\name{ModularExperiment}
\alias{ModularExperiment}
\title{ModularExperiment}
\usage{
ModularExperiment(assignments = character(), ...)
}
\description{
A class to represent the results of factor analysis
}
