% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R
\name{assignments}
\alias{assignments}
\alias{assignments<-}
\alias{assignments,ModularExperiment-method}
\alias{assignments<-,ModularExperiment-method}
\title{Get and set module feature assignments}
\usage{
\S4method{assignments}{ModularExperiment}(object, as_list = FALSE)

\S4method{assignments}{ModularExperiment}(object) <- value
}
\arguments{
\item{object}{\link[ReducedExperiment]{ModularExperiment} object.}

\item{as_list}{If \code{TRUE}, the results are returned as a list, with an entry
for each module containing a list of features.}

\item{value}{New value to replace existing assignments.}
}
\description{
Retrieves a vector of features (usually genes) named by the modules
they belong to.
}
