% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{module_preservation}
\alias{module_preservation}
\title{Get module preservation statistics}
\usage{
module_preservation(
  reference_dataset,
  test_dataset,
  reference_assay_name = "normal",
  test_assay_name = "normal",
  module_assignments = NULL,
  greyName = "module_0",
  goldName = "random",
  networkType = "signed",
  corFnc = "cor",
  ...
)
}
\arguments{
\item{reference_dataset}{The dataset that was used to define the modules.
Must be a \code{data.frame} or \code{matrix} with features as rows and samples as
columns, or a \link[ReducedExperiment]{ModularExperiment} or
\link[SummarizedExperiment]{SummarizedExperiment} object.}

\item{test_dataset}{The dataset that will be used to test for module
preservation. Must be a \code{data.frame} or \code{matrix} with features as rows and
samples as columns, or a \link[SummarizedExperiment]{SummarizedExperiment}
object. The features of \code{test_dataset} should be the same as
\code{reference_dataset}.}

\item{reference_assay_name}{If the reference dataset is a
\link[ReducedExperiment]{ModularExperiment} or
\link[SummarizedExperiment]{SummarizedExperiment} object, this argument
specifies which assay slot was used to define the modules.}

\item{test_assay_name}{If the reference dataset is a
\link[ReducedExperiment]{ModularExperiment} or
\link[SummarizedExperiment]{SummarizedExperiment} object, this argument
specifies which assay slot is to be used in preservation tests.}

\item{module_assignments}{If the reference dataset is not a
\link[ReducedExperiment]{ModularExperiment} object, this argument is
necessary to specify the module assignments.}

\item{greyName}{The name of the "module" of unassigned genes. Usually
"module_0" (ReducedExperiment) or "grey" (WGCNA).}

\item{goldName}{The name to be used for the "gold" module (which is made up
of a random sample of all network genes). Passed to
\link[WGCNA]{modulePreservation}.}

\item{networkType}{A string referring to the type of WGCNA network used for
the reference and test datasets. One of"unsigned", "signed" or
"signed hybrid". See \link[WGCNA]{adjacency}. Passed to
\link[WGCNA]{modulePreservation}.}

\item{corFnc}{A string referring to the function to be used to calculate
correlation. One of "cor" or "bicor". Passed to
\link[WGCNA]{modulePreservation}.}

\item{...}{Additional arguments to be passed to
\link[WGCNA]{modulePreservation}.}
}
\value{
A \code{data.frame} containing preservation statistics, as described
by \link[WGCNA]{modulePreservation}.
}
\description{
Tests whether a set of modules defined in the reference dataset are
preserved in the test dataset.
}
\examples{
# Get random matrices with rnorm, with 100 rows (features)
# and 20 columns (observations)
X_1 <- ReducedExperiment:::.makeRandomData(100, 20, "feature", "obs")
X_2 <- ReducedExperiment:::.makeRandomData(100, 20, "feature", "obs")

# Estimate 5 factors based on the data matrix
me_1 <- identify_modules(X, nc = 5, verbose = 0, powers = 5)
me_2 <- identify_modules(X, nc = 5, verbose = 0, powers = 5)

# Test module preservation
mp <- module_preservation(me_1, me_2, verbose = 10, nPermutations = 10)

# Plot the results
plot_module_preservation(mp)

}
