% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{plot_common_features}
\alias{plot_common_features}
\title{Heatmap comparing commonality across factors}
\usage{
plot_common_features(
  common_features,
  filename = NA,
  color = (grDevices::colorRampPalette(RColorBrewer::brewer.pal(n = 7, name =
    "YlOrRd")))(100)
)
}
\description{
Heatmap comparing commonality across factors
}
