% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{plot_module_preservation}
\alias{plot_module_preservation}
\title{Plot module preservation statistics}
\usage{
plot_module_preservation(
  module_preservation_results,
  show_random = TRUE,
  remove_module = NULL
)
}
\arguments{
\item{module_preservation_results}{The output of
\link[ReducedExperiment]{module_preservation}}

\item{show_random}{If TRUE, shows the random module in the plots.}

\item{remove_module}{The name of a module to be hidden from the plots.}
}
\value{
Two \code{ggplot2} plot objects combined by patchwork. Plots the
module preservation statistics generated by
\link[ReducedExperiment]{module_preservation}.
}
\description{
Plot module preservation statistics
}
\examples{
# Get random matrices with rnorm, with 100 rows (features)
# and 20 columns (observations)
X_1 <- ReducedExperiment:::.makeRandomData(100, 20, "feature", "obs")
X_2 <- ReducedExperiment:::.makeRandomData(100, 20, "feature", "obs")

# Estimate 5 factors based on the data matrix
me_1 <- identify_modules(X_1, nc = 5, verbose = 0, powers = 5)
me_2 <- identify_modules(X_2, nc = 5, verbose = 0, powers = 5)

# Test module preservation
mp <- module_preservation(me_1, me_2, verbose = 10, nPermutations = 10)

# Plot the results
plot_module_preservation(mp)

}
\author{
Jack Gisby
}
