% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ModularExperiment.R
\name{plotDendro,ModularExperiment-method}
\alias{plotDendro,ModularExperiment-method}
\title{Plot a dendrogram stored in a ModularExperiment}
\usage{
\S4method{plotDendro}{ModularExperiment}(
  object,
  groupLabels = "Module colors",
  dendroLabels = FALSE,
  hang = 0.03,
  addGuide = TRUE,
  guideHang = 0.05,
  color_func = WGCNA::labels2colors
)
}
\arguments{
\item{object}{\link[ReducedExperiment]{ModularExperiment} object.}

\item{groupLabels}{Module label axis label.
See \link[WGCNA]{plotDendroAndColors}.}

\item{hang}{See \link[WGCNA]{plotDendroAndColors}.}

\item{addGuide}{See \link[WGCNA]{plotDendroAndColors}.}

\item{guideHang}{See \link[WGCNA]{plotDendroAndColors}.}

\item{color_func}{Function for converting module names to colors.}

\item{If}{TRUE, shows feature names in the dendrogram.
See \link[WGCNA]{plotDendroAndColors}.}
}
\description{
Gets the dendrogram and plots it using
\link[WGCNA]{plotDendroAndColors}.
}
