% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R
\name{projectData,FactorisedExperiment,matrix-method}
\alias{projectData,FactorisedExperiment,matrix-method}
\alias{projectData,FactorisedExperiment,data.frame-method}
\alias{projectData,FactorisedExperiment,SummarizedExperiment-method}
\alias{predict,FactorisedExperiment-method}
\title{Project new data using pre-defined factors}
\usage{
\S4method{projectData}{FactorisedExperiment,matrix}(
  x,
  newdata,
  scale_reduced = TRUE,
  scale_newdata = NULL,
  center_newdata = NULL
)

\S4method{projectData}{FactorisedExperiment,data.frame}(
  x,
  newdata,
  scale_reduced = TRUE,
  scale_newdata = NULL,
  center_newdata = NULL
)

\S4method{projectData}{FactorisedExperiment,SummarizedExperiment}(
  x,
  newdata,
  scale_reduced = TRUE,
  scale_newdata = NULL,
  center_newdata = NULL,
  assay_name = "normal"
)

\S4method{predict}{FactorisedExperiment}(object, newdata, ...)
}
\arguments{
\item{newdata}{New data for projection. Must be a \code{data.frame} or \code{matrix}
with features as rows and samples as columns, or a
\link[SummarizedExperiment]{SummarizedExperiment} object.}

\item{scale_reduced}{Whether or not the reduced data should be scaled
after calculation.}

\item{scale_newdata}{Controls whether the \code{newdata} are scaled. If NULL,
performs scaling based on the \link[ReducedExperiment]{FactorisedExperiment}
object's \code{scale} slot. The value of this argument will be passed to the
\code{scale} argument of \link[base]{scale}.}

\item{center_newdata}{Controls whether the \code{newdata} are centered If NULL,
performs centering based on the \link[ReducedExperiment]{FactorisedExperiment}
object's \code{center} slot. The value of this argument will be passed to the
\code{center} argument of \link[base]{scale}.}

\item{assay_name}{If a \link[SummarizedExperiment]{SummarizedExperiment}
object is passed as new data, this argument indicates which assay should be
used for projection.}

\item{object}{A \link\link{FactorisedExperiment} object. The \code{loadings} slot of
this class will be used for projection. Additionally, by default, the \code{scale}
and \code{center} slots are used to apply the original transformation to the
new data.}
}
\value{
Calculates a matrix with samples as rows and factors as columns. If
\code{newdata} was a \code{matrix} or \code{data.frame}, this will be returned as a matrix.
If a \link[SummarizedExperiment]{SummarizedExperiment} object was passed
instead, then a If a \link[SummarizedExperiment]{FactorisedExperiment}
object will be created containing this matrix in its \code{reduced} slot.
}
\description{
Applies defined factors to new data.
}
