% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{reduced_gsea}
\alias{reduced_gsea}
\title{Gene set enrichment analysis}
\usage{
reduced_gsea(
  S,
  database = "msigdb_c2_cp",
  TERM2GENE = NULL,
  p_cutoff = 1,
  adj_method = "BH",
  nPermSimple = 1000,
  eps = 1e-10,
  ...
)
}
\description{
Performs \link[clusterProfiler]{GSEA} using the \code{clusterProfiler} package.
Expects a loadings matrix where features (genes) are rows and factors are
columns.
}
