% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_data.R
\name{run_ica}
\alias{run_ica}
\title{Run ICA}
\usage{
run_ica(
  X,
  nc,
  method = "imax",
  maxit = 500,
  tol = 1e-06,
  seed = 1,
  center_X = TRUE,
  scale_X = FALSE,
  reorient_skewed = TRUE,
  ...
)
}
\description{
Run ICA
}
