% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{run_ica}
\alias{run_ica}
\title{Run ICA for a data matrix}
\usage{
run_ica(
  X,
  nc,
  use_stability = FALSE,
  resample = FALSE,
  method = "fast",
  stability_threshold = NULL,
  center_X = TRUE,
  scale_X = FALSE,
  reorient_skewed = TRUE,
  seed = 1,
  scale_components = TRUE,
  scale_reduced = TRUE,
  n_runs = 30,
  BPPARAM = BiocParallel::SerialParam(),
  ...
)
}
\description{
Run ICA for a data matrix
}
