% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-ReducedExperiment.R
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames<-}
\alias{sampleNames,ReducedExperiment-method}
\alias{sampleNames<-,ReducedExperiment-method}
\title{Get sample names}
\usage{
\S4method{sampleNames}{ReducedExperiment}(object)

\S4method{sampleNames}{ReducedExperiment}(object) <- value
}
\arguments{
\item{object}{\link[ReducedExperiment]{ReducedExperiment} object.}

\item{value}{New value to replace existing names.}
}
\description{
Retrieves sample names (colnames).
}
