% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-FactorisedExperiment.R,
%   R/methods-ModularExperiment.R, R/methods-ReducedExperiment.R
\name{[,FactorisedExperiment,ANY,ANY,ANY-method}
\alias{[,FactorisedExperiment,ANY,ANY,ANY-method}
\alias{[,ModularExperiment,ANY,ANY,ANY-method}
\alias{[,ReducedExperiment,ANY,ANY,ANY-method}
\title{Extract and replace parts of ReducedExperiment objects}
\usage{
\S4method{[}{FactorisedExperiment,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)

\S4method{[}{ModularExperiment,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)

\S4method{[}{ReducedExperiment,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)
}
\arguments{
\item{x}{\link[ReducedExperiment]{ReducedExperiment} object.}

\item{i}{Slicing by rows (features, usually genes)}

\item{j}{Slicing by samples}

\item{k}{Slicing by reduced dimensions}

\item{...}{Additional arguments to be passed to the parent method.}

\item{drop}{Included for consistency with other slicing methods.}
}
\description{
Extract and replace parts of ReducedExperiment objects
}
